modified:   C++/1.2/01.cpp 	new file:   C++/1.2/02.cpp 	new file:   C++/1.2/03.cpp 	new file:   C++/1.2/04.cpp 	new file:   C++/1.2/05.cpp 	new file:   C++/1.2/06.cpp 	new file:   C++/1.2/07.cpp 	new file:   C++/1.2/08.cpp     new file:   C++/1.2/09.cpp     new file:   C++/1.2/10.cpp

+++ C++/1.2/02.cpp
@@ -2,7 +2,7 @@
 #include<iostream>
 using namespace std;
 main(){
-	cout<<sizeof(int)<<" "<<sizeof(short)<<endl;
+	cout<<sizeof(float)<<" "<<sizeof(double)<<endl;
 	system("pause");
 	return EXIT_SUCCESS;
 }
